Structured checkoutTest() that might replace checkout().

+++ src/main/java/com/example/models/Product.java
@@ -3,6 +3,7 @@ package com.example.models;
 import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.HashSet;
 import java.util.Set;
@@ -87,7 +88,7 @@ public class Product {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
